Conflicts:
	planet/Objects.ocd/HUD.ocd/FancyGridMenu.ocd/DefCore.txt
	planet/Objects.ocd/HUD.ocd/GridMenu.ocd/DefCore.txt
	planet/Objects.ocd/Helpers.ocd/Dummy.ocd/Graphics.png
	planet/Objects.ocd/Items.ocd/Tools.ocd/WindBag.ocd/Script.c
	planet/Objects.ocd/Items.ocd/Weapons.ocd/Javelin.ocd/Script.c
	planet/Objects.ocd/Libraries.ocd/ClonkControl.ocd/Script.c
	planet/Objects.ocd/Libraries.ocd/Constructor.ocd/ConstructionMenu.ocd/Script.c
	planet/Objects.ocd/Libraries.ocd/Constructor.ocd/Script.c
	planet/Objects.ocd/Libraries.ocd/Flag.ocd/Marker.ocd/Script.c
	planet/Objects.ocd/Libraries.ocd/Producer.ocd/Script.c
	planet/Objects.ocd/Structures.ocd/Flagpole.ocd/Script.c
	planet/Objects.ocd/Structures.ocd/Pump.ocd/Script.c
	src/game/C4Game.cpp
	src/gamescript/C4GameScript.cpp
	src/gui/C4GameMessage.cpp
	src/landscape/C4Landscape.cpp

+++ src/platform/C4SoundLoaders.cpp
@@ -131,10 +131,10 @@ int VorbisLoader::seek_func(void* datasource, ogg_int64_t offset, int whence)
 	switch (whence)
 	{
 	case SEEK_SET:
-		data->data_pos = offset < data->data_length ? offset : data->data_length;
+		data->data_pos = static_cast<size_t>(offset) < data->data_length ? static_cast<size_t>(offset) : data->data_length;
 		break;
 	case SEEK_CUR:
-		data->data_pos += offset < data->data_length - data->data_pos ? offset : data->data_length - data->data_pos;
+		data->data_pos += static_cast<size_t>(offset) < data->data_length - data->data_pos ? static_cast<size_t>(offset) : data->data_length - data->data_pos;
 		break;
 	case SEEK_END:
 		data->data_pos = data->data_length+1;
@@ -212,8 +212,6 @@ bool WavLoader::ReadInfo(SoundInfo* result, BYTE* data, size_t data_length, uint
 	ALuint wav = alutCreateBufferFromFileImage((const ALvoid *)data, data_length);
 	if (wav == AL_NONE)
 	{
-		// wouldn't want an error on any .ogg file
-		//LogF("load wav error: %s", alutGetErrorString(alutGetError()));
 		return false;
 	}
